+++ AppleRemote/GlobalKeyboardDevice.m
@@ -171,7 +171,7 @@ static OSStatus hotKeyEventHandler(EventHandlerCallRef inHandlerRef, EventRef in
 	EventHotKeyRef carbonHotKey;
 
 	hotKeyID.signature = 'PTHk';
-	hotKeyID.id = (long)keycode;
+	hotKeyID.id = (UInt32)keycode;
 	
 	err = RegisterEventHotKey(keycode, modifiers, hotKeyID, GetEventDispatcherTarget(), 0, &carbonHotKey );
