+++ FFTrack.h
@@ -107,7 +107,7 @@ extern AVPacket s_flushPacket;
 
 + (id)videoTrackWithAVStream:(AVStream*)stream index:(int)index;
 
-- (BOOL)initTrack:(int*)errorCode videoQueueCapacity:(int)videoQueueCapacity
+- (BOOL)initTrack:(int*)errorCode videoQueueCapacity:(NSInteger)videoQueueCapacity
                                      useFastDecoding:(BOOL)useFastDecoding;
 - (BOOL)setOpenGLContext:(NSOpenGLContext*)openGLContext
              pixelFormat:(NSOpenGLPixelFormat*)openGLPixelFormat
